deploy: add baseHref option to angular.json
When deploying to github pages, we need the baseHref to be
/cypher-companion/. When using the Git BASH interface, it is not
possible to specify '/cypher-companion/' on the command line without
triggering its MSYS path expansion. This can be worked around using
a specialized MSYS environment variable, 'MSYS_NO_PATHCONV=1'. However,
it is simpler to just specify this option in the config.

That also makes it as easy as running "ng deploy" in order to update the
github page!

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>