route: navigate away from summary automatically for large screens
When on small screens, we allow navigating separately between the
summary, and other views.

For larger displays, the summary is always visible and the link to jump
to the summary disappears. However, if you increase the size of the
screen width, we still show the summary in the router. Fix this by
automatically navigating away from the summary when we hit the small
breakpoint.

This ensures that we don't double up display of the summary component on
large screens.

In addition, automatically redirect the summary link to skills if the
user tries to navigate manually. Use a service to export the small
breakpoint instead of doing so manually.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>