use sessionStorage to save last view
Instead of storing the last view only during a single page load, store
it as a sessionStorage object. This way, we can use it even if the user
manually loads the summary page. Session storage is used so that we
don't remember this data forever, and only for a single browser session.

With this change, we now remember the last view when expanding the page
to split the summary and other views. This persists even when the user
has reloaded the page.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>