component: switch to d-flex for more control
We want the lower section of the display to completely fill the vertical
space. To make this possible, use d-flex to allow flex-grow, flex-fill,
etc. This is very tricky as we need to cascade the flexible fill and
flex display attributes all the way down into the final column divs. If
this is not done correctly, the resulting display will not behave as
expected.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>