app: add vertical bar between two components
Add a vertical bar between the two components when in wide display mode.
This ensures that the users can visually see the separation between the
two components.

Additionally, add a minimum width of 275px for the summary box. This
should be sufficient, and ensures that we never shrink below half the
size of the small breakpoint. Doing so makes it feel more natural when
the display size shrinks past the medium and into small breakpoints, as
we do not shrink the summary section too much until we completely run
out of space.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>