mainmenu: make routerLink activate tab
When navigating using the URL or a redirect, the ngbNav tabs do not get
activated properly. Fix this by adding routerLinkActive flags.

Reduce code duplication by building a link dictionary and using ngFor.
Wrap all links in divs, and conditionally add d-sm-none for the Summary
link that we hide in the large view display.

This change makes the links properly activate when redirecting and
ensures the correct tab activated both in small/modile view and in large
desktop view.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>