switch back to @angular/material and enable SCSS
The Drag and drop support for angular does not tie well into the
ng-bootstrap components. The main reason I switched to bootstrap was to
get some styling and support for the bootstrap grid layout.

I thought @angular/material didn't come with good styling, but this is
only if the project was not using SASS.

This change converts the project to use SCSS, rmeoves the ng-bootstrap
bits, and restores angular material. The change is large, but it is
difficult to do these steps without breaking functionality.

To make approximately equivalent behavior, the mainmenu component is
removed and pulled into the app.component.html. This is easier due to
how the divs are laid out.

Additionally, the nbg components are replaced by some material
components including a sidenav pop out for the small mobile friendly
menu.

This has slightly different behavior than before, but overall the same
functionality.

It will pave the way for using other angular material components which
work with features such as drag and drop.

The SCSS setup includes various bootstrap stylings in order to gain
advantage of the grid and container systems, as well as various utility
classes.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>