skills: add edit button and cleanup delete positioning
The existing delete button is placed in the mat-action-row which shows
up below the description and takes up unnecessary white space. Avoid
this by moving it into the normal description area.

Replace the icon with a mat-icon for consistency with other icons, since
the bootstrap icons don't seem to quite work as expected in all
locations.

Add an edit button and use the ms-auto bootstrap CSS class to put the
edit and delete icons to the right side of the screen.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>