skills: hook up skill addition
Convert the button toggle to a select, since this is the style expected
by form fields and allows hints. Update the save button to verify that
the skill name is set. Update the main skills component to call an
addSkill on the sheet to push the new skill to the end of the list.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>