skills: rename AddComponent to AddEditComponent
The AddComponent exists to allow adding a new skill to the skill list.
The logic is very similar to what we want for editing. Rather than
duplicating the functionality into a new component, we can re-use it by
converting the AddComponent into an AddEdit component.

First, rename the component from Add to AddEdit.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>