mainmenu: remove the background view
The original PDF includes both a "background" and "notes" section. This
is intended to allow saving the character background notes separate from
things such as campaign notes.

In the web view it doesn't really makes sense to have two separate tabs
for notes. Remove the background section. Players will be able to keep
separate notes with a title and a description, so they can hold the
background information in that area.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>