skills: implement edit functionality
Modify the AddEditComponent to support taking in a skill data, and
slightly modify its appearance for when editing skills.

This now gives us essentially full skill support by enabling add,
delete, edit, and re-ordering of skill data.

This will be the basis for implementing similar functionality across the
other components.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>