cyphers: enable add, edit, and delete for cyphers
Similar to abilities, allow adding and editing cyphers. Include support
for the cypher limit which will eventually be a part of the overall
summary of the character. Color the backgorund of cyphers past the
cypher limit with a light warning color. This helps indicate when the
user has too many cyphers. This is done instead of hard limiting because
some cypher systems allow going over the limit with consequences.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>