summary: display attribute pools
Split the attribute data into its own subclass and display the contents
through the summary tab.

To avoid code duplication, add an AttributePoolComponent to display the
attribute pool data.

The current display scheme isn't perfect, but its a close match to some
of the PDF layouts and should be suitable for now.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>