summary: implement pool increment and decrement
Add click events to the increment and decrement buttons enabling the
user to control the amount in the pools.

Logic to bound the increment and decrement between the range 0-maximum
is kept within the AttributePool class. We still use a separate
increment/decrement handler in the attribute-pool.component.ts file
because we will eventually want to emit a "save" event to save the
character sheet to the browser local storage. This can't really be done
from within the AttributePool class since its a sub-class of the
character sheet class.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>