summary: display name and character class information
Add a section displaying the character name, descriptor, type, focus,
and optional flavor.

Modify how the summary data is accessed to avoid many duplicate
variables. We end up making the summary component more tightly coupled
to the interface, but significantly reduce the number of accessors
needed.

Use *ngIf to more properly handle the case where the summary can't be
obtained, allowing use of a nullable type instead of the forced ':!'
used to assume its always defined.

Modify the cyphers component to avoid the need for a separate
cypherLimit accessor.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>