summary: add import/export buttons
Add buttons to import and export the character sheet data as a JSON
file.

Note that the import function is brittle. This is required because we
can't simply replace the references to objects. Instead, we must be
careful to ensure that all object refereces are updated in-place. Doing
this ensures that all views are updated consistently.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>