save and load character data via local storage
Use the HTML5 Local Storage interface to store the character data across
page loads. Use rxjs debounceTime and a Subject<> to debounce character
saves so that we limit saving to once every half a second. This also
ensures that saves minimize tearing by guaranteeing 500ms between change
requests from the user.

With this, the application should pretty much be functional. Character
sheet data should be saved between loads, and allow the user to import
and export in order to transfer saves between browsers.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>